Create open command
The primary motivator of creating this cli app was to have an easy way
to launch into any particular note file quickly with a series of easy
to remember commands. The `open` command will be used to pop open a
particular file in the editor of choice.

- Change config loader: we don't have any sort of default config anymore,
modules are responsible for bringing along their own defaults.
- Create `open` module with some starter functions to open up a file
by path in the editor of choice
- Begin to flesh out the api design for command modules/more config, shifting
away from a large case statement in the primary function