Use read-pkg module to load the version
The package.json file contains our module version, which is useful to display
for a command line utility. Previously I just pulled the value directly from
the json file, but it turns out this is problematic when compiling typescript
(as the file is now outside the project root).

The `read-pkg` module is simple and does the job well and avoids any of the
config or issues associated with trying to get around sourcing a file outside
of the project root.

- Add read-pkg module to index.ts
- Move a few of the arguments in the main function around
- Also add the daily/datepicker source code back in (was removed in an
  earlier commit while refactoring index)